Implement buffered sandbox factory
BufferedSBFactory maintains a buffer of started (then paused)
sandboxes created by a delegate factory. When Create is called,
it uses mount to make necessary files visible inside the sandbox.

Only work if mounting on the host side will be visible inside the
sandbox (i.e. delegate factory creates sandbox with slave volumes).

`sandbox_buffer` in the JSON config defines the capacity of the
buffer. If it is 0, non-buffered sandbox factory will be used.