handler: add containers to HandlerManager after pause
Pausing a container takes ~50us with olcontainers and ~10ms with Docker.
This discrepency leads to unintended reuse due to it being on the
HandlerManager's queue for longer with Docker than with olcontainers.
To even the playing field, add for reuse only after it has been paused.
This is only a concern when the handler lru size is set to 0.