handler: only add to the LRU if the handler fits in the cache
There exists a scheduling condition when using the LRU for eviction.
Instead, the handler should be deleted if it is the last reference and
does not fit in the LRU. This releases the locks and reacquires them in
the same order as the Get() call to make sure we are the last instance
of the handler.