Fix Docker container usage with named pipes.
There was a race condition where the worker would attempt to read
from the pipe before it was created in handler.go. To solve this,
we create it if it doesn't exist and append if it does on both ends.

Also added a new "spin" process which will run inside of Docker
containers instead of ol-init, which now does some named-pipe
notification which we don't need with Docker.