Install packages from the host
Import cache entries and their child handler interpreters need to
share a package directory to function properly for packages that
import "lazily." To achieve this, we bind mount a read-only copy
of a single shared package directory where all installs are made.

This commit implements this, removing the installation requirements
from the python servers and moving them to an "installManager" on
the host.